Fix duration of Part B left hand
Those notes aren’t held for the duration of a quarter note – they’re
held for the duration of an *eighth* note (plus perhaps the tiniest bit
more: tenuto), so that they *contrast* with the right hand. I’m not sure
if that’s how it’s notated for the brass section in the original march,
but for piano I’m convinced that this is the right way to write it.

I decided against turning the non-tenuto eights into sixteenths plus
rests, though: I think that much can be left to the pianist (though the
MIDI output makes it clear that these should not be held for quite their
full duration either).